refactor: 상수 관리하는 패키지 파일 생성 및 이동
