+++ src/main/java/bridge/Constants/CommandConstant.java
@@ -1,4 +1,4 @@
-package bridge.domain;
+package bridge.Constants;
 
 public class CommandConstant {
     private CommandConstant() {
