+++ src/main/java/bridge/Constants/ErrorMessageConstant.java
@@ -1,4 +1,4 @@
-package bridge.domain;
+package bridge.Constants;
 
 public final class ErrorMessageConstant {
