+++ src/test/java/bridge/service/BridgeMakerTest.java
@@ -1,6 +1,6 @@
 package bridge.service;
 
-import bridge.domain.ErrorMessageConstant;
+import bridge.Constants.ErrorMessageConstant;
 import bridge.util.BridgeNumberGenerator;
 import bridge.util.BridgeRandomNumberGenerator;
 import org.junit.jupiter.api.Test;
