+++ src/test/java/bridge/service/InputHandlerImplTest.java
@@ -1,9 +1,7 @@
 package bridge.service;
 
-import bridge.domain.ErrorMessageConstant;
+import bridge.Constants.ErrorMessageConstant;
 import org.junit.jupiter.api.Test;
-
-import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
